Introducing flow_solver_cases subdirectory and adapting FlowSolver class for 1D problems.
