+++ src/testing/CMakeLists.txt
@@ -20,7 +20,8 @@ set(TESTS_SOURCE
     optimization_inverse_manufactured/optimization_inverse_manufactured.cpp
     burgers_rewienski_snapshot.cpp
     reduced_order.cpp
-    flow_solver.cpp)
+    flow_solver.cpp
+    flow_solver_cases/periodic_cube_flow.cpp)
 
 foreach(dim RANGE 1 3)
     # Output library
